flask_backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.114.2
Flask==3.0.3
h11==0.14.0
idna==3.9
itsdangerous==2.2.0
Jinja2==3.1.4
line_profiler==4.1.3
MarkupSafe==2.1.5
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
sniffio==1.3.1
spotipy==2.24.0
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
Werkzeug==3.0.4
